Changed to double

+++ src/math.h
@@ -8,4 +8,5 @@
  *
  */
 
-float curve(float p0, float p1, float p2, float p3, float t);
+//float curve(float p0, float p1, float p2, float p3, float t);
+double curve(double p0, double p1, double p2, double p3, double t);
